requirements.txt
appdirs==1.4.3
appnope==0.1.0
astroid==1.4.9
autopep8==1.3.2
beautifulsoup4==4.5.3
bleach==2.0.0
boto3==1.4.4
botocore==1.5.54
click==6.7
cloudpickle==0.2.2
dask==0.14.3
decorator==4.0.11
dill==0.2.5
distributed==1.16.3
docutils==0.13.1
entrypoints==0.2.2
enum34==1.1.6
findspark==1.1.0
fn==0.4.3
future==0.16.0
gapic-google-cloud-datastore-v1==0.15.3
gapic-google-cloud-error-reporting-v1beta1==0.15.3
gapic-google-cloud-logging-v2==0.91.3
gapic-google-cloud-pubsub-v1==0.15.3
gapic-google-cloud-spanner-admin-database-v1==0.15.3
gapic-google-cloud-spanner-admin-instance-v1==0.15.3
gapic-google-cloud-spanner-v1==0.15.3
gapic-google-cloud-speech-v1beta1==0.15.3
gapic-google-cloud-vision-v1==0.90.3
google-auth==0.9.0
google-auth-httplib2==0.0.2
google-cloud==0.23.0
google-cloud-bigquery==0.23.0
google-cloud-bigtable==0.23.1
google-cloud-core==0.23.1
google-cloud-datastore==0.23.0
google-cloud-dns==0.23.0
google-cloud-error-reporting==0.23.1
google-cloud-language==0.23.1
google-cloud-logging==0.23.1
google-cloud-monitoring==0.23.0
google-cloud-pubsub==0.23.0
google-cloud-resource-manager==0.23.0
google-cloud-runtimeconfig==0.23.0
google-cloud-spanner==0.23.1
google-cloud-speech==0.23.0
google-cloud-storage==0.23.1
google-cloud-translate==0.23.0
google-cloud-vision==0.23.3
google-gax==0.15.8
googleapis-common-protos==1.5.2
grpc-google-iam-v1==0.11.1
grpcio==1.2.0
hdf5storage==0.1.14
HeapDict==1.0.0
html5lib==0.999999999
httplib2==0.10.3
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
isort==4.2.5
jedi==0.10.2
Jinja2==2.9.6
jmespath==0.9.2
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.1.0
jupyter-core==4.3.0
lazy-object-proxy==1.2.2
locket==0.2.0
lxml==3.7.3
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.7.4
mock==2.0.0
msgpack-python==0.4.8
nbconvert==5.1.1
nbformat==4.3.0
notebook==5.0.0
numexpr==2.6.2
numpy==1.12.0
oauth2client==3.0.0
packaging==16.8
pandas==0.19.2
pandocfilters==1.4.1
partd==0.3.8
pbr==3.0.0
pexpect==4.2.1
pickleshare==0.7.4
ply==3.8
prompt-toolkit==1.0.14
proto-google-cloud-datastore-v1==0.90.3
proto-google-cloud-error-reporting-v1beta1==0.15.3
proto-google-cloud-logging-v2==0.91.3
proto-google-cloud-pubsub-v1==0.15.3
proto-google-cloud-spanner-admin-database-v1==0.15.3
proto-google-cloud-spanner-admin-instance-v1==0.15.3
proto-google-cloud-spanner-v1==0.15.3
proto-google-cloud-speech-v1beta1==0.15.3
proto-google-cloud-vision-v1==0.90.3
protobuf==3.2.0
psutil==5.2.2
ptyprocess==0.5.1
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycodestyle==2.3.1
PyFunctional==1.0.0
Pygments==2.2.0
pylint==1.6.5
pyparsing==2.2.0
pyperclip==1.5.27
python-dateutil==2.6.0
pytz==2016.10
pyzmq==16.0.2
qtconsole==4.3.0
requests==2.13.0
rsa==3.4.2
s3fs==0.1.0
s3transfer==0.1.10
simplegeneric==0.8.1
six==1.10.0
sortedcollections==0.5.3
sortedcontainers==1.5.7
tables==3.4.2
tabulate==0.7.7
tblib==1.3.2
terminado==0.6
testpath==0.3
toolz==0.8.2
toree==0.1.0
tornado==4.5.1
tqdm==4.11.2
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
wget==3.2
widgetsnbextension==2.0.0
wrapt==1.10.10
xmltodict==0.11.0
yappi==0.98
zict==0.1.2
